pyproject.toml
[project]
name = "osac-test-infra"
version = "0.1.0"
description = "OSAC end-to-end test infrastructure"
requires-python = ">=3.10"
dependencies = [
    "pytest>=8.0.0",
    "pytest-bdd>=7.0.0",
    "httpx>=0.27.0",
    "kubernetes>=29.0.0",
    "pyyaml>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
bdd_features_base_dir = "features"
